// frontend/src/lib/api/client/types.gen.ts
export type TaskAction = 'START' | 'INFO' | 'WARN' | 'SUCCESS' | 'ERROR' | 'AI_ENRICHED' | 'GEOCODED' | 'ROUTED';

/**
 * TaskLogRead
 *
 * Schema for reading a TaskLog.
 */
export type TaskLogRead = {
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at?: string | null;
    /**
     * Id
     */
    id: string;
    /**
     * Task Id
     */
    task_id: string;
    action: TaskAction;
    /**
     * Message
     */
    message: string;
    /**
     * Data
     */
    data?: {
        [key: string]: unknown;
    } | null;
};